requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
langchain==1.2.13
langchain-google-genai==4.2.1
langchain-anthropic==1.4.0
langchain-community==0.4.1
chromadb==1.5.5
python-dotenv==1.2.2
